Update heapsort.c
Fixed the prompt message to be clearer: "Enter the number of elements" instead of "Enter no of elements".

Adjusted the spacing in the print statements to improve readability.

Replaced the main() function signature with int main(void). Although using void main() is commonly seen, it is not a recommended practice according to the C standard.